:fire: Remove old files and packages
- Remove chakra-UI
- Remove Dashboard
- Profile
- Restructure configs and files
