+++ package.json
@@ -4,29 +4,23 @@
   "private": true,
   "dependencies": {
     "@ant-design/icons": "^4.2.2",
-    "@chakra-ui/core": "^0.8.0",
     "@craco/craco": "^5.7.0",
-    "@emotion/core": "^10.0.35",
-    "@emotion/styled": "^10.0.27",
     "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.0.4",
-    "@testing-library/user-event": "^12.1.7",
-    "@types/jest": "^26.0.14",
-    "@types/node": "^14.11.2",
-    "@types/react": "^16.9.50",
+    "@testing-library/react": "^11.1.0",
+    "@testing-library/user-event": "^12.1.10",
+    "@types/jest": "^26.0.15",
+    "@types/node": "^14.14.2",
+    "@types/react": "^16.9.53",
     "@types/react-dom": "^16.9.8",
-    "antd": "^4.6.6",
+    "antd": "^4.7.2",
     "axios": "^0.20.0",
     "craco-less": "^1.17.0",
-    "emotion-theming": "^10.0.27",
-    "formik": "^2.1.7",
     "lodash": "^4.17.20",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3",
-    "typescript": "~4.0.3",
-    "yup": "^0.29.3"
+    "react-scripts": "3.4.4",
+    "typescript": "~4.0.3"
   },
   "scripts": {
     "start": "BROWSER=none craco start",
@@ -52,21 +46,21 @@
     ]
   },
   "devDependencies": {
-    "@types/lodash": "^4.14.161",
-    "@types/react-router-dom": "^5.1.5",
-    "@types/react-select": "^3.0.21",
-    "@types/yup": "^0.29.7",
-    "@typescript-eslint/eslint-plugin": "^4.3.0",
-    "@typescript-eslint/parser": "^4.3.0",
-    "cypress": "^5.3.0",
-    "eslint": "^7.10.0",
+    "@types/lodash": "^4.14.162",
+    "@types/react-router-dom": "^5.1.6",
+    "@types/react-select": "^3.0.22",
+    "@types/yup": "^0.29.8",
+    "@typescript-eslint/eslint-plugin": "^4.5.0",
+    "@typescript-eslint/parser": "^4.5.0",
+    "cypress": "^5.4.0",
+    "eslint": "^7.11.0",
     "eslint-config-airbnb": "^18.2.0",
-    "eslint-config-prettier": "^6.12.0",
+    "eslint-config-prettier": "^6.14.0",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsx-a11y": "^6.3.1",
     "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.21.3",
-    "eslint-plugin-react-hooks": "^4.1.2",
+    "eslint-plugin-react": "^7.21.5",
+    "eslint-plugin-react-hooks": "^4.2.0",
     "husky": ">=4",
     "lint-staged": ">=10",
     "prettier": "^2.1.2"
