+++ src/components/Navbar.tsx
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import { Menu, Typography } from 'antd';
 import { Link, useLocation } from 'react-router-dom';
 import _ from 'lodash';
-import { getItem } from '../utils/sessionStorage';
-import { configData, menuNames } from '../utils/configs';
+
+import { menuNames } from '../utils/configs';
 
 const menusInitial = [menuNames.signIn, menuNames.signUp];
 const menusForLoggedIn = [menuNames.dash, menuNames.profile, menuNames.signOut];
@@ -11,11 +11,10 @@ const menusForLoggedIn = [menuNames.dash, menuNames.profile, menuNames.signOut];
 const Navbar = (): JSX.Element => {
   const location = useLocation();
   const { Title } = Typography;
-  const [accessData] = useState(getItem(configData.accessTokenKeyName));
   const [menus, setMenus] = useState([...menusInitial]);
 
   const renderMenus = () => {
-    const isLoggedIn = _.isEmpty(accessData);
+    const isLoggedIn = _.isEmpty({});
     if (!isLoggedIn) {
       setMenus([...menusForLoggedIn]);
     }
