+++ src/configs/types/common.ts
