+++ src/configs/types/navlink.ts
