--- src/interfaces/signin.ts
@@ -1,10 +1,0 @@
-import { HandleCommonFunc } from '../types/common';
-
-export interface SignIn {
-  email: string;
-  password: string;
-}
-
-export interface SignInProps {
-  handleAccessData: HandleCommonFunc;
-}
