--- src/interfaces/signout.ts
@@ -1,6 +1,0 @@
-import { HandleCommonFunc, AccessDataObject } from '../types/common';
-
-export interface SignOutProps {
-  accessData: AccessDataObject;
-  handleAccessData: HandleCommonFunc;
-}
