--- src/interfaces/signup.ts
@@ -1,6 +1,0 @@
-export interface SignUpProps {
-  name: string;
-  email: string;
-  password: string;
-  passwordConfirmation: string;
-}
